requirements.txt
Flask==2.1.0
flask-cors==3.0.10
Pillow==9.0.1
ultralytics==0.1.28
gunicorn==21.2.0